pyproject.toml
[tool.poetry]
name = "pythonproject2"
version = "0.1.0"
description = ""
authors = ["arcimiendar <artimendor@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.1.2"
SQLAlchemy = "^1.4.32"
python-telegram-bot = "^13.13"

[tool.poetry.dev-dependencies]
sqlalchemy2-stubs = "^0.0.2-alpha.24"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
